+++ src/common/CResourceBase.h
@@ -371,7 +371,7 @@ public:
 
 	bool IsLinked() const;	// been loaded from the scripts ?
 	CResourceScript * GetLinkFile() const;
-	long GetLinkOffset() const;
+	int GetLinkOffset() const;
 	void SetLink( CResourceScript * pScript );
     void CopyTransfer( CResourceLink * pLink );
 	void ScanSection( RES_TYPE restype );
