+++ src/graysvr/CPathFinder.h
@@ -119,7 +119,7 @@ protected:
 
 protected:
 	void Clear();
-	unsigned long Heuristic(CPathFinderPointRef& Pt1,CPathFinderPointRef& Pt2);
+	unsigned int Heuristic(CPathFinderPointRef& Pt1,CPathFinderPointRef& Pt2);
 	void GetChildren(CPathFinderPointRef& Point, std::list<CPathFinderPointRef>& ChildrenRefList );
 	void FillMap();	// prepares map with walkable statuses
 };
