+++ src/graysvr/CWorld.h
@@ -192,9 +192,9 @@ public:
 	unsigned int m_SectorsQty;
 
 public:
-	long m_iSaveCountID;		// Current archival backup id. Whole World must have this same stage id
-	long m_iLoadVersion;		// Previous load version. (only used during load of course)
-	long m_iPrevBuild;			// Previous __GITREVISION__
+	int m_iSaveCountID;			// Current archival backup id. Whole World must have this same stage id
+	int m_iLoadVersion;			// Previous load version. (only used during load of course)
+	int m_iPrevBuild;			// Previous __GITREVISION__
 	CServTime m_timeStartup;	// When did the system restore load/save ?
 
 	CGrayUID m_uidLastNewItem;	// for script access.
